fix test_vulnerability_detector
