+++ api/app/tests/unittests/test_vulnerability_detector.py
@@ -1,5 +1,4 @@
 import pytest
-from sqlalchemy.orm import Session
 
 from app import models
 from app.detector import vulnerability_detector
@@ -12,9 +11,10 @@ class TestGetAffectedVersions:
         # Given
         affect = models.Affect(
             vuln_id="dummy_vuln_id",
-            package_id="dummy_package_id",
             affected_versions=["1.0.0"],
             fixed_versions=["2.0.0"],
+            affected_name="dummy_package_name",
+            ecosystem="dummy_package_ecosystem",
         )
 
         # When
@@ -39,9 +39,10 @@ class TestGetAffectedVersions:
         # Given
         affect = models.Affect(
             vuln_id="dummy_vuln_id",
-            package_id="dummy_package_id",
             affected_versions=[],
             fixed_versions=fixed_versions,
+            affected_name="dummy_package_name",
+            ecosystem="dummy_package_ecosystem",
         )
 
         # When
@@ -65,9 +66,10 @@ class TestGetAffectedVersions:
         # Given
         affect = models.Affect(
             vuln_id="dummy_vuln_id",
-            package_id="dummy_package_id",
             affected_versions=[],
             fixed_versions=fixed_versions,
+            affected_name="dummy_package_name",
+            ecosystem="dummy_package_ecosystem",
         )
 
         # When
@@ -93,9 +95,10 @@ class TestCheckMatchedPackageVersionAndAffect:
         )
         affect = models.Affect(
             vuln_id="dummy_vuln_id",
-            package_id=package.package_id,
             affected_versions=[">=1.0.0"],
             fixed_versions=[],
+            affected_name=package.name,
+            ecosystem=package.ecosystem,
         )
 
         # When
@@ -121,9 +124,10 @@ class TestCheckMatchedPackageVersionAndAffect:
         )
         affect = models.Affect(
             vuln_id="dummy_vuln_id",
-            package_id=package.package_id,
             affected_versions=[">=1.1.0"],
             fixed_versions=[],
+            affected_name=package.name,
+            ecosystem=package.ecosystem,
         )
 
         # When
@@ -149,9 +153,10 @@ class TestCheckMatchedPackageVersionAndAffect:
         )
         affect = models.Affect(
             vuln_id="dummy_vuln_id",
-            package_id=package.package_id,
             affected_versions=[">=1.0.0"],
             fixed_versions=[],
+            affected_name=package.name,
+            ecosystem=package.ecosystem,
         )
 
         # When
@@ -177,9 +182,10 @@ class TestCheckMatchedPackageVersionAndAffect:
         )
         affect = models.Affect(
             vuln_id="dummy_vuln_id",
-            package_id=package.package_id,
             affected_versions=["NG"],
             fixed_versions=[],
+            affected_name=package.name,
+            ecosystem=package.ecosystem,
         )
 
         # When
@@ -209,9 +215,10 @@ class TestCheckMatchedPackageVersionAndAffect:
         )
         affect = models.Affect(
             vuln_id="dummy_vuln_id",
-            package_id=package2.package_id,
             affected_versions=[">=1.0.0"],
             fixed_versions=[],
+            affected_name=package2.name,
+            ecosystem=package2.ecosystem,
         )
 
         # When
@@ -221,48 +228,3 @@ class TestCheckMatchedPackageVersionAndAffect:
 
         # Then
         assert not matched
-
-
-class TestDetectVulnerabilityByAffect:
-    def test_it_should_return_matched_package_version_ids_when_affected_versions_match(
-        self,
-        testdb: Session,
-    ):
-        # Given
-        package = models.Package(
-            name="test_package",
-            ecosystem="npm",
-        )
-        package_version_1 = models.PackageVersion(
-            version="1.1.0",
-            package_id=package.package_id,
-            package=package,
-        )
-        package_version_2 = models.PackageVersion(
-            version="1.2.0",
-            package_id=package.package_id,
-            package=package,
-        )
-        package_version_3 = models.PackageVersion(
-            version="1.3.0",
-            package_id=package.package_id,
-            package=package,
-        )
-        package.package_versions = [package_version_1, package_version_2, package_version_3]
-        affect = models.Affect(
-            vuln_id="dummy_vuln_id",
-            package_id=package.package_id,
-            affected_versions=["<=1.2.0"],
-            fixed_versions=[],
-            package=package,
-        )
-
-        # When
-        matched_package_version_ids = vulnerability_detector.detect_vulnerability_by_affect(
-            testdb, affect
-        )
-
-        # Then
-        assert package_version_1.package_version_id in matched_package_version_ids
-        assert package_version_2.package_version_id in matched_package_version_ids
-        assert package_version_3.package_version_id not in matched_package_version_ids
